Fix propeller axis definition

+++ tools/run_optimization.py
@@ -41,9 +41,9 @@ parser.add_argument('--mesh', type=str, default='L2')
 # Case information
 parser.add_argument('--cruiseset', type=str, default='cruiseset_00.py')
 parser.add_argument('--rotation', type=str, default='1')
-parser.add_argument('--prop_x', type=int, default=-0.2)
-parser.add_argument('--prop_y', type=int, default=0.0)
-parser.add_argument('--prop_z', type=int, default=0.3)
+parser.add_argument('--prop_x', type=float, default=-0.2)
+parser.add_argument('--prop_y', type=float, default=0.0)
+parser.add_argument('--prop_z', type=float, default=0.3)
 # Geometric information
 parser.add_argument('--ffd', type=str, default='10x10')
 parser.add_argument('--disksurf', type=str, default='disk_surf')
